Check for font changes when changing states
Panels need to be repositioned if font changes. This happened when
exiting the font selection dialog, but it did not when exiting the
preferences dialog, which can be used to change the font also.
Putting the check in the push/pop state handlers removes the need
for each state to deal with repositioning after changing fonts.