Save panel locations as multiples of font size
Previously it stored the integer multiple + additional pixel offset,
but adding that additional pixel offset defeats the purpose. Now the
location is stored as a float for x and a float for y. Files that
were previously made, that have ints for x and y, will still be read
correctly.

This is how a1558fe should be implemented from the start, and makes
8aa9012 obsolete.