Change zoom calculation and limit to 10 ticks
cf94f7b was a stupid change that changed the zoom to 20 ticks, but it
only made the ticks between zoom levels smaller instead of moving the
limit of how much one can zoom out.

This commit reverts that and tackles the actual zooming function so
it zooms out more.